Adds the secret Santa assignment function
Function generate_pairings takes in a list of emails formatted as
['a@a.com', 'b@b.com', 'c@c.com'] and returns a similar list
containing the assigned partner in the same order. For example:
if a@a.com is assigned to c@c.com, the first item in the returned
list is c@c.com.

The algorithm works by shuffling the list and daisy chaining the
pairings. It does a check for preventing someone from getting
themselves as their own secret Santa. It also prevents sub-cycles
like A->B and B->A.

This algorithm is inspired by a chapter in Dr. Hannah Fry's book
"The Indisputable Existence of Santa Claus"